Move looper rendering to looper module (#76)
